requirements.txt
alembic==1.12.0
asgiref==3.6.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.8.0
dj-rest-auth==5.0.2
Django==4.1.7
django-allauth==0.61.1
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-email-verification==0.3.3
django-environ==0.11.2
django-filter==23.5
django-google-fonts==0.0.3
django-htmx==1.14.0
django-mathfilters==1.0.0
django-taggit==3.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.4.2
email-validator==2.0.0.post2
factory-boy==3.2.1
Faker==17.6.0
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
netaddr==0.9.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
pluggy==1.3.0
protobuf==4.21.12
psycopg2==2.9.9
pycparser==2.21
pygame==2.5.0
PyJWT==1.7.1
PyMySQL==1.1.0
pytest==7.4.2
pytest-html==4.0.2
pytest-json==0.4.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.10.0
SQLAlchemy==2.0.21
sqlparse==0.4.3
stripe==7.6.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2022.7
urllib3==2.1.0
validators==0.22.0
webencodings==0.5.1
Werkzeug==2.3.0
wrapt==1.16.0
WTForms==3.0.1
yookassa==2.5.0
zipp==3.17.0
